+++ src/entities/contacts/contacts.module.scss
@@ -26,28 +26,17 @@
 
 .container {
 	display: flex;
+	// justify-content: center;
 	align-items: center;
 }
 
 .input_extra {
 	width: 100%;
 	display: block;
-	padding: 12px 36px 12px 12px;
+	// padding: 12px 36px 12px 12px;
 	margin: 0;
 	border-radius: 12px;
 	min-height: 50px;
-	box-sizing: border-box;
-	border: 1px solid;
-	color: $--Foreground-default;
-	background-color: $--bg-primary;
-	font-size: 16px;
-	font-style: normal;
-	font-weight: 400;
-	line-height: 23px;
-	position: relative;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	overflow: hidden;
 
 	&::placeholder {
 		color: $--Foreground-subtitle;
@@ -78,6 +67,7 @@
 	height: 50px;
 	padding: 0px 10px;
 	width: auto;
+	border-color:  $--Foreground-default;
 	border-radius: 12px;
 
 	&::placeholder {
