+++ src/services/ProjectService.ts
@@ -1,4 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+import { ProjectService } from './models/IProject';
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 
 export const projectsApi = createApi({
@@ -26,7 +27,7 @@ export const projectsApi = createApi({
 			}),
 		}),
 		postProject: builder.mutation({
-			query: ({ body }) => ({
+			query: (body : ProjectService) => ({
 				url: `/projects/`,
 				method: 'POST',
 				body: body,
