+++ src/services/models/IProject.ts
@@ -4,3 +4,28 @@ export interface IProject {
 	image: string;
 	isLiked: boolean;
 }
+
+
+export type ProjectService = {
+	id: number,
+	started: string;
+	ended: string;
+	name: string;
+	directions: [
+		{
+			id: number;
+			name: string;
+		},
+	];
+	project_specialists: [
+		{
+			id: number;
+			profession: {
+				id: number;
+				specialization: string;
+				specialty: string;
+			};
+		},
+	];
+
+}
